fix: int64 indexing migration batch 6 - overflow checks and loop counters
- NdArrayToJaggedArray.cs: Add overflow checks for managed array limits,
  explicit (int) casts with validation before allocation, loop comparisons
  against .Length instead of shape[x]
- NDArray.matrix_power.cs: Add overflow check for np.eye dimension
- NDArray.Indexing.Masking.cs: Fix loop counter int→long for mask.size iteration
- Update INT64_MIGRATION_PROGRESS.md with session 5 fixes and audit results
  (np.load.cs, np.save.cs confirmed as valid exceptions)

+++ src/NumSharp.Core/Selection/NDArray.Indexing.Masking.cs
@@ -259,8 +259,8 @@ namespace NumSharp
             var result = new NDArray(this.dtype, new Shape(resultShape));
 
             // Copy selected slices
-            int destIdx = 0;
-            for (int srcIdx = 0; srcIdx < mask.size; srcIdx++)
+            long destIdx = 0;
+            for (long srcIdx = 0; srcIdx < mask.size; srcIdx++)
             {
                 if (mask.GetBoolean(srcIdx))
                 {
